Add a way to skip `Valid::{check, batch_check}` when checks are trivial
