+++ serialize/src/impls/mod.rs
@@ -1,0 +1,4 @@
+mod collections;
+mod int_like;
+mod misc;
+mod tuples;
