+++ serialize/src/lib.rs
@@ -6,7 +6,7 @@
     rust_2018_idioms,
     rust_2021_compatibility
 )]
-#![forbid(unsafe_code)]
+#![deny(unsafe_code)]
 #![doc = include_str!("../README.md")]
 mod error;
 mod flags;
@@ -68,26 +68,44 @@ pub enum Validate {
 }
 
 pub trait Valid: Sync {
+    /// Whether the `check` method is trivial (i.e. always returns `Ok(())`). If this is `true`,
+    /// the `batch_check` method will skip all checks and return `Ok(())`.
+    /// This should be set to `true` for types where `check` is trivial, e.g.
+    /// integers, field elements, etc.
+    /// This is `false` by default.
+    /// This is primarily an optimization to skip unnecessary checks in `batch_check`.
+    const TRIVIAL_CHECK: bool = false;
+
+    /// Checks whether `self` is valid. If `self` is valid, returns `Ok(())`. Otherwise, returns
+    /// an error describing the failure.
+    /// This method is called by `deserialize_with_mode` if `validate` is `Validate::Yes`.
     fn check(&self) -> Result<(), SerializationError>;
 
+    /// Checks whether all items in `batch` are valid. If all items are valid, returns `Ok(())`.
+    /// Otherwise, returns an error describing the first failure.
+    #[inline]
     fn batch_check<'a>(
         batch: impl Iterator<Item = &'a Self> + Send,
     ) -> Result<(), SerializationError>
     where
         Self: 'a,
     {
-        #[cfg(feature = "parallel")]
-        {
-            use rayon::{iter::ParallelBridge, prelude::ParallelIterator};
-            batch.par_bridge().try_for_each(|e| e.check())?;
-        }
-        #[cfg(not(feature = "parallel"))]
-        {
-            for item in batch {
-                item.check()?;
+        if Self::TRIVIAL_CHECK {
+            Ok(())
+        } else {
+            #[cfg(feature = "parallel")]
+            {
+                use rayon::{iter::ParallelBridge, prelude::ParallelIterator};
+                batch.par_bridge().try_for_each(|e| e.check())?;
             }
+            #[cfg(not(feature = "parallel"))]
+            {
+                for item in batch {
+                    item.check()?;
+                }
+            }
+            Ok(())
         }
-        Ok(())
     }
 }
 
@@ -115,20 +133,25 @@ pub trait CanonicalSerialize {
         compress: Compress,
     ) -> Result<(), SerializationError>;
 
+    /// Returns the size in bytes of the serialized version of `self` with the given compression mode.
     fn serialized_size(&self, compress: Compress) -> usize;
 
+    /// Serializes `self` into `writer` using the compressed form if applicable.
     fn serialize_compressed<W: Write>(&self, writer: W) -> Result<(), SerializationError> {
         self.serialize_with_mode(writer, Compress::Yes)
     }
 
+    /// Returns the size in bytes of the compressed serialized version of `self`.
     fn compressed_size(&self) -> usize {
         self.serialized_size(Compress::Yes)
     }
 
+    /// Serializes `self` into `writer` using the uncompressed form.
     fn serialize_uncompressed<W: Write>(&self, writer: W) -> Result<(), SerializationError> {
         self.serialize_with_mode(writer, Compress::No)
     }
 
+    /// Returns the size in bytes of the uncompressed serialized version of `self`.
     fn uncompressed_size(&self) -> usize {
         self.serialized_size(Compress::No)
     }
@@ -158,18 +181,41 @@ pub trait CanonicalDeserialize: Valid + Sized {
         validate: Validate,
     ) -> Result<Self, SerializationError>;
 
+    /// Reads `Self` from `reader` using the compressed form if applicable.
+    /// Performs validation if applicable.
     fn deserialize_compressed<R: Read>(reader: R) -> Result<Self, SerializationError> {
         Self::deserialize_with_mode(reader, Compress::Yes, Validate::Yes)
     }
 
+    /// Reads `Self` from `reader` using the compressed form if applicable, without validating the
+    /// deserialized value.
+    ///
+    /// # Note
+    ///
+    /// This should be used with caution, as it may lead to invalid values being used in
+    /// subsequent computations, which may have security implications depending on the context.
+    ///
+    /// It should only be used when the caller can guarantee that the input is valid, e.g. if it was
+    /// generated by a trusted source.
     fn deserialize_compressed_unchecked<R: Read>(reader: R) -> Result<Self, SerializationError> {
         Self::deserialize_with_mode(reader, Compress::Yes, Validate::No)
     }
 
+    /// Reads `Self` from `reader` using the uncompressed form. Performs validation if applicable.
     fn deserialize_uncompressed<R: Read>(reader: R) -> Result<Self, SerializationError> {
         Self::deserialize_with_mode(reader, Compress::No, Validate::Yes)
     }
 
+    /// Reads `Self` from `reader` using the uncompressed form, without validating the deserialized
+    /// value.
+    ///
+    /// # Note
+    ///
+    /// This should be used with caution, as it may lead to invalid values being used in
+    /// subsequent computations, which may have security implications depending on the context.
+    ///
+    /// It should only be used when the caller can guarantee that the input is valid, e.g. if it was
+    /// generated by a trusted source.
     fn deserialize_uncompressed_unchecked<R: Read>(reader: R) -> Result<Self, SerializationError> {
         Self::deserialize_with_mode(reader, Compress::No, Validate::No)
     }
@@ -214,8 +260,10 @@ impl<H: Digest> ark_std::io::Write for HashMarshaller<'_, H> {
     }
 }
 
-/// The CanonicalSerialize induces a natural way to hash the
-/// corresponding value, of which this is the convenience trait.
+/// [`CanonicalSerialize`] induces a natural way to hash a value:
+/// serialize it and hash the resulting byte string.
+///
+/// This is a convenience trait that combines the two steps.
 pub trait CanonicalSerializeHashExt: CanonicalSerialize {
     fn hash<H: Digest>(&self) -> GenericArray<u8, <H as OutputSizeUser>::OutputSize> {
         let mut hasher = H::new();
@@ -232,8 +280,9 @@ pub trait CanonicalSerializeHashExt: CanonicalSerialize {
     }
 }
 
-/// CanonicalSerializeHashExt is a (blanket) extension trait of
-/// CanonicalSerialize
+/// [`CanonicalSerializeHashExt`] is a (blanket) extension trait of
+/// `CanonicalSerialize`: all types implementing the latter
+/// automatically implement the former.
 impl<T: CanonicalSerialize> CanonicalSerializeHashExt for T {}
 
 #[inline]
